Add files via upload
While these files are barely modified from the ROS server/client tutorial, they still allowed me to determine the nature of the delay between the maximum time taken for the information to be requested by the client and then heard back from the server. I believe this method is correct however I was unable to properly configure my ROS setup to get a unique self-made server/client setup to run.