Writeup
These plots attempt to show the time taken for a ROS publisher and subscriber to share information, as well as for a client and server. I strongly believe that while the method I employed to find this duration worked for the publisher/subscriber, I do not believe it was accurate for the client/server. This is due to two reasons. The first being that the duration for the server/client should be less than that of the publisher subscriber on average. In addition, the durations for the client feedback do not follow any sort of Gaussian trend, indicating that the method I employed to determine these times was likely incorrect.